Update to reflect our most novel and important projects (#3)
## What is the goal of this PR?

Some of the data in this repository is a bit behind where we are now, and people keep adding stars! We want people to get an accurate picture of the current state of the TypeDB ecosystem from this repo, so we have updated the resources to reflect this.

## What are the changes implemented in this PR?

We bring some of the most exciting content forward. Previously we've linked just to https://github.com/vaticle/typedb-examples or https://github.com/typedb-osi, but now we display links to our most exciting content immediately.